+++ tests/rules/mod.rs
@@ -1,0 +1,1 @@
+pub mod typing_any_used;
